# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Warehouse Management - API Documentation",
        "description": "Warehouse Management open api documentation",
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Main api documentation"
        }
    ],
    "paths": {
...
        "/Item/GetItems": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "get all items",
                "operationId": "a5c88234b46ff6dd90543c5fe658c34f",
                "responses": {
                    "200": {
                        "description": "Successful Operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "status",
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "description": "data",
                                            "type": "object"
                                        },
                                        "message": {
                                            "description": "message",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
        "/Item/InsertItem": {
            "post": {
                "tags": [
                    "Item"
                ],
                "summary": "Create a new item",
                "operationId": "ce287b3d9ae5e8e996f85f26dbe96a2f",
                "requestBody": {
                    "description": "Body request needed to create a new item",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "productId": {
                                        "description": "productId"
                                    },
                                    "imei": {
                                        "description": "imei"
                                    },
                                    "name": {
                                        "description": "name"
                                    },
                                    "companyId": {
                                        "description": "companyId"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "status",
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "description": "data",
                                            "type": "object"
                                        },
                                        "message": {
                                            "description": "message",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "status",
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "description": "data",
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "message": {
                                            "description": "message",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "status",
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "description": "data",
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "message": {
                                            "description": "message",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerToken": []
                    }
                ]
            }
        },
        "/Item/UpdateItem": {
            "post": {
                "tags": [
                    "Item"
                ],
                "summary": "Update item",
                "operationId": "d49d85a7098aed9950b058b8f8798aca",
                "requestBody": {
                    "description": "Body request needed to update a item",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    },
                                    "productId": {
                                        "description": "productId"
                                    },
                                    "imei": {
                                        "description": "imei"
                                    },
                                    "name": {
                                        "description": "name"
                                    },
                                    "companyId": {
                                        "description": "companyId"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "status",
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "description": "data",
                                            "type": "object"
                                        },
                                        "message": {
                                            "description": "message",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "status",
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "description": "data",
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "message": {
                                            "description": "message",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "status",
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "description": "data",
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "message": {
                                            "description": "message",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerToken": []
                    }
                ]
            }
        },
        "/Item/DeleteItem": {
            "delete": {
                "tags": [
                    "Item"
                ],
                "summary": "Delete Item",
                "operationId": "d23c8e9264e0ec7173122c03b2c25aa3",
                "requestBody": {
                    "description": "Body request needed to delete item",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "status",
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "description": "data",
                                            "type": "object"
                                        },
                                        "message": {
                                            "description": "message",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "status",
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "description": "data",
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "message": {
                                            "description": "message",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "status",
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "description": "data",
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "message": {
                                            "description": "message",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerToken": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerToken": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}